Refactors

+++ flopp-kernel/src/main/java/com/github/kjetilv/flopp/kernel/MemorySegments.java
@@ -50,7 +50,7 @@ public final class MemorySegments {
     public static long bytesAt(MemorySegment memorySegment, long offset, long count) {
         long data = 0;
         long lastPosition = offset + count - 1;
-        for (long i = 0; i < count; i++) {
+        for (int i = 0; i < count; i++) {
             byte b = memorySegment.get(JAVA_BYTE, lastPosition - i);
             data = (data << ALIGNMENT) + b;
         }
